Added comment documentation to classes #1186
