+++ src/PSRule.Benchmark/PSRule.Benchmark.csproj
@@ -7,6 +7,7 @@
     <ProjectGuid>{3ec0912f-bfc7-4b53-a1a1-0ba993c6282e}</ProjectGuid>
     <EnableNuget>false</EnableNuget>
     <IsPackable>false</IsPackable>
+    <GenerateDocumentationFile>false</GenerateDocumentationFile>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)'=='Release'">
