+++ src/PSRule/Data/TargetIssueInfo.cs
@@ -1,35 +1,55 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using System;
 using System.Management.Automation;
 using Newtonsoft.Json;
 
 namespace PSRule.Data
 {
-    public sealed class TargetIssueInfo
+    /// <summary>
+    /// An issue reported by a downstream tool.
+    /// </summary>
+    public sealed class TargetIssueInfo : IEquatable<TargetIssueInfo>
     {
         private const string PROPERTY_TYPE = "type";
         private const string PROPERTY_NAME = "name";
         private const string PROPERTY_PATH = "path";
         private const string PROPERTY_MESSAGE = "message";
 
+        /// <summary>
+        /// Create an empty issue.
+        /// </summary>
         public TargetIssueInfo()
         {
             // Do nothing
         }
 
+        /// <summary>
+        /// The type of issue.
+        /// </summary>
         [JsonProperty(PropertyName = PROPERTY_TYPE)]
         public string Type { get; internal set; }
 
+        /// <summary>
+        /// The name of the issue.
+        /// </summary>
         [JsonProperty(PropertyName = PROPERTY_NAME)]
         public string Name { get; internal set; }
 
+        /// <summary>
+        /// The object path reported by the issue.
+        /// </summary>
         [JsonProperty(PropertyName = PROPERTY_PATH)]
         public string Path { get; internal set; }
 
+        /// <summary>
+        /// A localized message describing the issue.
+        /// </summary>
         [JsonProperty(PropertyName = PROPERTY_MESSAGE)]
         public string Message { get; internal set; }
 
+        /// <inheritdoc/>
         public bool Equals(TargetIssueInfo other)
         {
             return other != null &&
@@ -39,11 +59,13 @@ namespace PSRule.Data
                 Message == other.Message;
         }
 
+        /// <inheritdoc/>
         public override bool Equals(object obj)
         {
             return obj is TargetIssueInfo info && Equals(info);
         }
 
+        /// <inheritdoc/>
         public override int GetHashCode()
         {
             unchecked // Overflow is fine
@@ -57,11 +79,17 @@ namespace PSRule.Data
             }
         }
 
+        /// <summary>
+        /// Create an issue from a structured object.
+        /// </summary>
         public static TargetIssueInfo Create(object o)
         {
             return o is PSObject pso ? Create(pso) : null;
         }
 
+        /// <summary>
+        /// Create an issue from a structured object.
+        /// </summary>
         public static TargetIssueInfo Create(PSObject o)
         {
             var result = new TargetIssueInfo();
