+++ src/PSRule/Definitions/SourceExtent.cs
@@ -3,12 +3,24 @@
 
 namespace PSRule.Definitions
 {
+    /// <summary>
+    /// A source location for a PSRule expression.
+    /// </summary>
     public interface ISourceExtent
     {
+        /// <summary>
+        /// The source file path.
+        /// </summary>
         string File { get; }
 
+        /// <summary>
+        /// The first line of the expression.
+        /// </summary>
         int? Line { get; }
 
+        /// <summary>
+        /// The first position of the expression.
+        /// </summary>
         int? Position { get; }
     }
