+++ src/PSRule/Runtime/Operand.cs
@@ -5,27 +5,60 @@ using System;
 
 namespace PSRule.Runtime
 {
+    /// <summary>
+    /// The type of operand that is compared with the expression.
+    /// </summary>
     public enum OperandKind
     {
+        /// <summary>
+        /// Unknown.
+        /// </summary>
         None = 0,
 
+        /// <summary>
+        /// An object path.
+        /// </summary>
         Path = 1,
 
+        /// <summary>
+        /// The object target type.
+        /// </summary>
         Type = 2,
 
+        /// <summary>
+        /// The object target name.
+        /// </summary>
         Name = 3,
 
+        /// <summary>
+        /// The object source information.
+        /// </summary>
         Source = 4,
 
+        /// <summary>
+        /// The target object itself.
+        /// </summary>
         Target = 5
     }
 
+    /// <summary>
+    /// An operand that is compared with PSRule expressions.
+    /// </summary>
     public interface IOperand
     {
+        /// <summary>
+        /// The value of the operand.
+        /// </summary>
         object Value { get; }
 
+        /// <summary>
+        /// The type of operand.
+        /// </summary>
         OperandKind Kind { get; }
 
+        /// <summary>
+        /// The object path to the operand.
+        /// </summary>
         string Path { get; }
     }
